app/requirements.txt
fastapi
uvicorn
jinja2
ics
python-multipart
python-dateutil

beautifulsoup4>=4.9.3
python-jose[cryptography]
passlib[bcrypt]
email-validator